pyproject.toml
[project]
name = "fastapi-fullauth"
version = "0.1.0"
description = "Production-grade, async-native authentication and authorization library for FastAPI"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["fastapi", "authentication", "authorization", "security", "jwt", "webauthn"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.110",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "pyjwt[cryptography]>=2.8",
    "argon2-cffi>=23.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
sqlalchemy = ["sqlalchemy[asyncio]>=2.0", "alembic>=1.13"]
sqlmodel = ["sqlmodel>=0.0.16", "alembic>=1.13"]
tortoise = ["tortoise-orm>=0.21"]
beanie = ["beanie>=1.25"]
redis = ["redis>=5.0"]
passkeys = ["webauthn>=2.0"]
totp = ["pyotp>=2.9", "qrcode>=7.4"]
oauth = ["httpx-oauth>=0.13"]
magic-link = ["itsdangerous>=2.1"]
audit = []
tenants = []
all = [
    "sqlalchemy[asyncio]>=2.0", "alembic>=1.13",
    "sqlmodel>=0.0.16",
    "tortoise-orm>=0.21",
    "beanie>=1.25",
    "redis>=5.0",
    "webauthn>=2.0",
    "pyotp>=2.9", "qrcode>=7.4",
    "httpx-oauth>=0.13",
    "itsdangerous>=2.1",
]

[project.urls]
Homepage = "https://github.com/mdfarhankc/fastapi-fullauth"
Documentation = "https://github.com/mdfarhankc/fastapi-fullauth"
Repository = "https://github.com/mdfarhankc/fastapi-fullauth"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fastapi_fullauth"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["B008", "B904"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
    "uvicorn>=0.42.0",
]
